[Add] Bookmark Swagger 작성 완료 [ userId 임시 값 처리]

+++ back/src/bookmark/bookmark.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { PharmacyBookMark } from '@prisma/client';
 @Injectable()
@@ -8,34 +8,41 @@ export class BookmarkService {
     userId: string,
     pharmacyId: number,
   ): Promise<PharmacyBookMark> {
-    return this.prisma.pharmacyBookMark.create({
+    return await this.prisma.pharmacyBookMark.create({
       data: {
         user_id: userId,
         pharmacy_id: pharmacyId,
       },
     });
   }
-  async deleteBookmark(id: number): Promise<PharmacyBookMark> {
-    return this.prisma.pharmacyBookMark.delete({
+  async deleteBookmark(id: number): Promise<void> {
+    const bookmark = await this.getBookmark(id);
+    if (!bookmark) {
+      throw new ForbiddenException(`Bookmark with id ${id} not found`);
+    }
+    await this.prisma.pharmacyBookMark.delete({
       where: {
         id: id,
       },
     });
   }
 
-  async listBookmark(userId): Promise<PharmacyBookMark[]> {
-    return this.prisma.pharmacyBookMark.findMany({
+  async listBookmark(): Promise<PharmacyBookMark[]> {
+    return await this.prisma.pharmacyBookMark.findMany({
       where: {
-        user_id: userId,
+        user_id: '7fdd64c5-0ef1-49ef-abdc-55f44e048aa5',
+      },
+      include: {
+        Pharmacy: true,
       },
     });
   }
 
-  async getBookmark(userId, pharmacyId: number): Promise<PharmacyBookMark> {
-    return this.prisma.pharmacyBookMark.findFirst({
+  async getBookmark(id: number): Promise<PharmacyBookMark> {
+    return await this.prisma.pharmacyBookMark.findFirst({
       where: {
-        user_id: userId,
-        pharmacy_id: pharmacyId,
+        user_id: '7fdd64c5-0ef1-49ef-abdc-55f44e048aa5',
+        id,
       },
       include: {
         Pharmacy: true,
